+++ packages/module/src/common/helpers/index.ts
@@ -1,4 +1,2 @@
 export * from './generateNonce';
-export * from './mergeSignatures';
 export * from './readPlutusData';
-export * from './readTransaction';
