+++ packages/module/src/index.ts
@@ -1,10 +1,6 @@
 export * from './common/contracts';
 export * from './common/types';
-export {
-  generateNonce,
-  readPlutusData,
-  readTransaction,
-} from './common/helpers';
+export { generateNonce, readPlutusData } from './common/helpers';
 export * from './common/utils/resolver';
 export * from './common/utils/parser';
 export * from './core/CIP2';
