+++ packages/module/src/serializer/index.ts
@@ -1,0 +1,1 @@
+export * from './signatures';
